Remove itemRenderer from Colors.
